remove old eloplay application
