+++ src/components/navigation-bar.tsx
@@ -5,13 +5,13 @@ export default function NavigationBar() {
   return (
     <div className="fixed bottom-0 right-0 flex w-full justify-center">
       <div className="flex w-[min(500px,100%)] justify-around bg-background-secondary">
-        <Link href={"/new/leaderboard"} className="p-4">
+        <Link href={"/leaderboard"} className="p-4">
           <Medal className="text-primary" size={30} />
         </Link>
-        <Link href={"/new/addLeagueMatch"} className="p-4">
+        <Link href={"/addLeagueMatch"} className="p-4">
           <PlusCircle className="text-primary" size={30} />
         </Link>
-        <Link href={"/new/leagueActivity"} className="p-4">
+        <Link href={"/leagueActivity"} className="p-4">
           <HeartPulse className="text-primary" size={30} />
         </Link>
       </div>
