+++ src/pages/team/join/[id].tsx
