+++ src/pages/teamUser/[id].tsx
